requirements.txt
altair                       5.3.0
annotated-types              0.7.0
appnope                      0.1.4
asttokens                    2.4.1
attrs                        23.2.0
blinker                      1.8.2
cachetools                   5.3.3
certifi                      2024.6.2
charset-normalizer           3.3.2
click                        8.1.7
comm                         0.2.2
debugpy                      1.8.2
decorator                    5.1.1
docx                         0.2.4
exceptiongroup               1.2.1
executing                    2.0.1
gitdb                        4.0.11
GitPython                    3.1.43
google-ai-generativelanguage 0.6.6
google-api-core              2.19.1
google-api-python-client     2.135.0
google-auth                  2.31.0
google-auth-httplib2         0.2.0
google-generativeai          0.7.1
googleapis-common-protos     1.63.2
grpcio                       1.64.1
grpcio-status                1.62.2
httplib2                     0.22.0
idna                         3.7
ipykernel                    6.29.5
ipython                      8.26.0
jedi                         0.19.1
Jinja2                       3.1.4
jsonschema                   4.22.0
jsonschema-specifications    2023.12.1
jupyter_client               8.6.2
jupyter_core                 5.7.2
lxml                         5.2.2
markdown-it-py               3.0.0
MarkupSafe                   2.1.5
matplotlib-inline            0.1.7
mdurl                        0.1.2
nest-asyncio                 1.6.0
numpy                        2.0.0
packaging                    24.1
pandas                       2.2.2
parso                        0.8.4
pexpect                      4.9.0
pillow                       10.4.0
pip                          24.1.1
platformdirs                 4.2.2
plotly                       5.22.0
prompt_toolkit               3.0.47
proto-plus                   1.24.0
protobuf                     4.25.3
psutil                       6.0.0
ptyprocess                   0.7.0
pure-eval                    0.2.2
pyarrow                      16.1.0
pyasn1                       0.6.0
pyasn1_modules               0.4.0
pydantic                     2.8.0
pydantic_core                2.20.0
pydeck                       0.9.1
Pygments                     2.18.0
pyparsing                    3.1.2
PyPDF2                       3.0.1
python-dateutil              2.9.0.post0
python-docx                  1.1.2
python-dotenv                1.0.1
pytz                         2024.1
pyzmq                        26.0.3
referencing                  0.35.1
requests                     2.32.3
rich                         13.7.1
rpds-py                      0.18.1
rsa                          4.9
setuptools                   63.2.0
six                          1.16.0
smmap                        5.0.1
stack-data                   0.6.3
streamlit                    1.36.0
tenacity                     8.4.2
toml                         0.10.2
toolz                        0.12.1
tornado                      6.4.1
tqdm                         4.66.4
traitlets                    5.14.3
typing_extensions            4.12.2
tzdata                       2024.1
uritemplate                  4.1.1
urllib3                      2.2.2
wcwidth                      0.2.13